Add scheduler to maintain webtasks.
The main purpose of this script is to handle the process of triggering
different webtasks. Every webtask that is triggered gets data about TFA
usage from particular service like Github, Slack etc.

Another purpose of scheduler is to handle notification process which is
delegated to the different webtask.

This structure helps to extend scheduler easily by adding new webtasks
(services) and notifiers like Twitter, Slack, HipChat and so on.